refactor: move discord logic to different classes
